Order : Refactored ID datatype to long

+++ src/main/java/com/kuna_backend/dtos/order/OrderDto.java
@@ -4,8 +4,8 @@ import com.kuna_backend.models.Order;
 import org.jetbrains.annotations.NotNull;
 
 public class OrderDto {
-    private Integer id;
-    private @NotNull Integer clientId;
+    private long id;
+    private @NotNull long clientId;
 
     public OrderDto() {
     }
@@ -14,19 +14,19 @@ public class OrderDto {
         this.setId(order.getId());
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public Integer getClientId() {
+    public long getClientId() {
         return clientId;
     }
 
-    public void setClientId(Integer clientId) {
+    public void setClientId(long clientId) {
         this.clientId = clientId;
     }
 }
